+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 CHANGELOG
 ==========
 
+- 2026-04-13 WsPanel: Persist selected topics using local storage
 - 2026-04-13 WsPanel: Show message stores in topics list
 - 2026-04-13 RestPanel: Update stop and abort mapping button styles
 - 2026-04-13 RestPanel: Fix REST API error display and styling